Moved Batch Size Managers into a separate file
Also renamed clip arguments as they were shadowing built-in functions
