+++ bs_scheduler/__init__.py
@@ -1,12 +1,10 @@
+from .batch_size_manager import BatchSizeManager, DefaultBatchSizeManager, CustomBatchSizeManager
 from .batch_size_schedulers import LambdaBS, MultiplicativeBS, StepBS, MultiStepBS, ConstantBS, LinearBS, ExponentialBS, \
     SequentialBS, PolynomialBS, CosineAnnealingBS, ChainedBSScheduler, IncreaseBSOnPlateau, CyclicBS, \
-    CosineAnnealingBSWithWarmRestarts, OneCycleBS, BSScheduler, BatchSizeManager
+    CosineAnnealingBSWithWarmRestarts, OneCycleBS, BSScheduler
 
-# We do not export DefaultBatchSizeManager and CustomBatchSizeManager because they are not needed. Users with custom
-# setups can create their own batch size managers.
 
 __all__ = ['LambdaBS', 'MultiplicativeBS', 'StepBS', 'MultiStepBS', 'ConstantBS', 'LinearBS', 'ExponentialBS',
            'SequentialBS', 'PolynomialBS', 'CosineAnnealingBS', 'ChainedBSScheduler', 'IncreaseBSOnPlateau', 'CyclicBS',
-           'CosineAnnealingBSWithWarmRestarts', 'OneCycleBS', 'BSScheduler', 'BatchSizeManager']
-
-del batch_size_schedulers  # noqa: F821
+           'CosineAnnealingBSWithWarmRestarts', 'OneCycleBS', 'BSScheduler', 'BatchSizeManager',
+           'DefaultBatchSizeManager', 'CustomBatchSizeManager']
